Fixed some issues and add update order logic.

+++ src/containers/CompanyDetail/CompanyDetailScreen.tsx
@@ -9,7 +9,8 @@ import {authReset} from '../../redux/slices/authSlice';
 import {wp} from '../../styles/responsiveScreen';
 import {RootScreens} from '../../types/type';
 import {mediumFont} from '../../styles';
-import {Platform} from 'react-native';
+import {useLogoutMutation} from '../../api/auth';
+import utils from '../../helper/utils';
 
 const CompanyDetailScreen = ({navigation, route}: any) => {
   const from = route.params.from;
@@ -17,6 +18,7 @@ const CompanyDetailScreen = ({navigation, route}: any) => {
   const dispatch = useDispatch();
   const [loading, setLoading] = useState(false);
   const [isOpen, setIsOpen] = useState(false);
+  const [logout, {isLoading}] = useLogoutMutation();
 
   React.useLayoutEffect(() => {
     if (from === 'Profile') {
@@ -48,13 +50,18 @@ const CompanyDetailScreen = ({navigation, route}: any) => {
 
   const logoutPress = async () => {
     setIsOpen(false);
-    setLoading(true);
-    await dispatch(authReset());
-    await AsyncStorage.clear();
-    const keysToRemove = ['token', 'MyCart', 'MyAddressList', 'NotiToken'];
-    await AsyncStorage.multiRemove(keysToRemove);
-    setLoading(false);
-    resetNavigateTo(navigation, RootScreens.Login);
+    const {data, error}: any = await logout();
+    if(!error && data.statusCode === 200) {
+      setLoading(true);
+      await dispatch(authReset());
+      await AsyncStorage.clear();
+      const keysToRemove = ['token', 'MyCart', 'MyAddressList', 'NotiToken'];
+      await AsyncStorage.multiRemove(keysToRemove);
+      setLoading(false);
+      resetNavigateTo(navigation, RootScreens.Login);
+    } else {
+      utils.showErrorToast(error.data.message);
+    }
   };
 
   return (
@@ -62,7 +69,7 @@ const CompanyDetailScreen = ({navigation, route}: any) => {
       <CompanyDetail
         from={from}
         navigation={navigation}
-        loading={loading}
+        loading={loading || isLoading}
         loginData={data}
       />
       <Popup
